replace reflection in test with package-private initForTest methods
Move testECDHSM2ManualMatchesBC to com.jcraft.jsch.bc package so it
can call package-private initForTest() on ECDHSM2/ECDHSM2BC directly,
eliminating the setAccessible() calls that fail the forbiddenapis check.
