+++ src/main/java/com/jcraft/jsch/bc/ECDHSM2BC.java
@@ -80,6 +80,12 @@ public class ECDHSM2BC implements ECDH {
     Q_array = pubKey.getQ().getEncoded(false);
   }
 
+  /** For testing only: inject a pre-generated key pair instead of generating one randomly. */
+  void initForTest(ECPrivateKeyParameters priv, byte[] publicPoint) {
+    privateKey = priv;
+    Q_array = publicPoint;
+  }
+
   @Override
   public byte[] getQ() throws Exception {
     return Q_array;
